- Added a hovering effect when moving the mouse cursor over an item in the user interface - Added a function to add an entire inventory list to the shop's inventory - If the client is launched without a running backend, a message pops up - Ad tool-tip over the SellIn and Quality columns shows the numeric value of these properties - Updated README.md - Fixed namespaces - Fixed progression of backstage passes
