+++ GildedRose.sln
@@ -7,7 +7,9 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GildedRose.Client", "Gilded
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GildedRose.Server", "GildedRose.Server\GildedRose.Server.csproj", "{568E3050-2260-48D5-A0C0-7E36AEECE248}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GildedRose.Contracts", "GildedRose.Contracts\GildedRose.Contracts.csproj", "{08FAE514-64B9-4306-809F-3518665AE0BB}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GildedRose.Contracts", "GildedRose.Contracts\GildedRose.Contracts.csproj", "{08FAE514-64B9-4306-809F-3518665AE0BB}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GildedRose.DataSource", "GildedRose.DataSource\GildedRose.DataSource.csproj", "{8D622AA2-50E3-42A8-896B-65E4980819F8}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -27,6 +29,10 @@ Global
 		{08FAE514-64B9-4306-809F-3518665AE0BB}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{08FAE514-64B9-4306-809F-3518665AE0BB}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{08FAE514-64B9-4306-809F-3518665AE0BB}.Release|Any CPU.Build.0 = Release|Any CPU
+		{8D622AA2-50E3-42A8-896B-65E4980819F8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{8D622AA2-50E3-42A8-896B-65E4980819F8}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{8D622AA2-50E3-42A8-896B-65E4980819F8}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{8D622AA2-50E3-42A8-896B-65E4980819F8}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
